code review changes - moved addNewAuthorizedOauthToConfig to oauth.js and renamed to authenticateWithOauth and updated references
