+++ packages/cms-lib/oauth.js
@@ -1,7 +1,7 @@
 const OAuth2Manager = require('@hubspot/api-auth-lib/OAuth2Manager');
 const { updatePortalConfig, getPortalConfig } = require('./lib/config');
 const { logger, logErrorInstance } = require('./logger');
-const { AUTH_METHODS } = require('./lib/constants');
+const { AUTH_METHODS, DEFAULT_OAUTH_SCOPES } = require('./lib/constants');
 
 const oauthManagers = new Map();
 
@@ -52,14 +52,23 @@ const addOauthToPortalConfig = (portalId, oauth) => {
       portalId,
     });
     logger.log('Configuration updated');
-    process.exit();
   } catch (err) {
     logErrorInstance(err);
   }
 };
 
+const authenticateWithOauth = async configData => {
+  const portalId = parseInt(configData.portalId, 10);
+  const oauth = setupOauth(portalId, {
+    scopes: DEFAULT_OAUTH_SCOPES,
+    ...configData,
+  });
+  logger.log('Authorizing');
+  await oauth.authorize();
+  addOauthToPortalConfig(portalId, oauth);
+};
+
 module.exports = {
   getOauthManager,
-  setupOauth,
-  addOauthToPortalConfig,
+  authenticateWithOauth,
 };
